stage2: kvm: arm64: Save/get esr_el2 in shadow_ctxt

+++ arch/arm64/include/asm/hypsec_vcpu.h
@@ -40,7 +40,7 @@ void __save_shadow_kvm_regs(struct kvm_vcpu *vcpu, u64 ec);
 void __restore_shadow_kvm_regs(struct kvm_vcpu *vcpu);
 
 void update_exception_gp_regs(struct kvm_vcpu *vcpu);
-extern int sec_el2_handle_sys_reg(struct kvm_vcpu *vcpu, unsigned long esr);
+extern int sec_el2_handle_sys_reg(struct kvm_vcpu *vcpu, u32 esr);
 
 void __save_encrypted_vcpu(struct kvm_vcpu *vcpu);
 void __update_exception_shadow_flag(struct kvm_vcpu *vcpu, int exp);
